requirements.txt
Flask==3.0.3
openai==1.28.1
openai_whisper==20231117
